backend/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
aiohttp==3.9.1

# OpenAI
openai==1.12.0

# Supabase
supabase==2.3.4

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
python-slugify==8.0.1
waybackpy==3.0.6
duckduckgo-search==4.4.3
python-Wappalyzer==0.3.1
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Caching & Rate Limiting
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Type hints
typing-extensions==4.9.0
